update userAuthCheck test
